+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# @blsq/blsq-report-components@1.0.112
+
+- [contracts] add possibility to cache calculations during contract validations
+- [invoices] add a page /reports/:period/:orgUnitId to allow a fixed point link with other apps (iaso, hesabu manager, orbf2)
+
+# @blsq/blsq-report-components@1.0.110
+
+- [dataentry] allow data entries over multiple datasets mixing [multiple](https://github.com/BLSQ/blsq-report-components/pull/186) periodicity
+- [completeness] adapt the completeness logic to match the data entry adaptations
+
+# @blsq/blsq-report-components@1.0.110
+
+- [contracts] displays disabled orgunit name in dialog from contract show page
+- [contracts] displays column with invalid contracts warnings, creates filter for invalid contracts, sorts invalid contracts first by default
+
+# @blsq/blsq-report-components@1.0.109
+
+- [invoice] allow to calculate simple indicators in the mapper
+- [approvals] creates module for managing data approvals
+
+# @blsq/blsq-report-components@1.0.108
+
+- [invoice] allow to hide print button and make default the invoice prompt to empty if no invoice
+
+# @blsq/blsq-report-components@1.0.107
+
+- [dataentry] allow to specify fullWidth on Dhis2Input component
+
 # @blsq/blsq-report-components@1.0.106
 
 - [contracts] add 2 new stats block about contract/sub contracts